cheri: Fix compiler warnings
Fix CHERI related compiler warnings by using proper cast operators.

+++ include/linux/rbtree.h
@@ -31,9 +31,9 @@
 
 /* 'empty' nodes are nodes that are known not to be inserted in an rbtree */
 #define RB_EMPTY_NODE(node)  \
-	((node)->__rb_parent_color == (unsigned long)(node))
+	((node)->__rb_parent_color == (uintptr_t)(node))
 #define RB_CLEAR_NODE(node)  \
-	((node)->__rb_parent_color = (unsigned long)(node))
+	((node)->__rb_parent_color = (uintptr_t)(node))
 
 
 extern void rb_insert_color(struct rb_node *, struct rb_root *);
@@ -59,7 +59,7 @@ extern void rb_replace_node_rcu(struct rb_node *victim, struct rb_node *new,
 static inline void rb_link_node(struct rb_node *node, struct rb_node *parent,
 				struct rb_node **rb_link)
 {
-	node->__rb_parent_color = (unsigned long)parent;
+	node->__rb_parent_color = (uintptr_t)parent;
 	node->rb_left = node->rb_right = NULL;
 
 	*rb_link = node;
@@ -68,7 +68,7 @@ static inline void rb_link_node(struct rb_node *node, struct rb_node *parent,
 static inline void rb_link_node_rcu(struct rb_node *node, struct rb_node *parent,
 				    struct rb_node **rb_link)
 {
-	node->__rb_parent_color = (unsigned long)parent;
+	node->__rb_parent_color = (uintptr_t)parent;
 	node->rb_left = node->rb_right = NULL;
 
 	rcu_assign_pointer(*rb_link, node);
